Started Text-to-Speech branch which may work on Hololens.
The project contains two empty game objects. Audio Manager contains TextToSpeech.cs which provides functionality for converting text to a Unity audio source. Root contains a test script which makes calls to TextToSpeech.cs and should run on startup. The project does not run successfully in Unity editor due to the preprocessor directives in TextToSpeech.cs but may run in the Hololens due to it being UWP so should have the Windows namespace.