Updated Scripts in Unity Project
Program now successfully works in Unity. Uses TTS to prompt the user to speak, takes voice input, repeats what was spoken by the user, converts this place name to coordinates and converts these coordinates to a route object. Not built or tested outside of Unity as I don't have the capabilities to run Mixed Reality or the Hololens Emulator.